Update CRUD UI tests and Corresponding Step Definition
The updates have been made to both the Cucumber scenarios for the CRUD UI and the corresponding step definitions. The layout of CRUD UI feature file has been modified for better readability and organization of scenarios. Also, the labels @crud @ui have been moved in the right scenario for User Edit a Post.

The modifications in the step definition file involve changes that improve the test assertion for checking if a post already exists with known content. Now, instead of throwing a pending exception, an actual implementation for the step has been provided. Other steps have been updated to follow a similar format as well.

This is going to improve the overall reliability of our functional tests.