Separate test scenarios between UI and API
Deleted the unnecessary configuration file browsers.json since we have moved to use Docker-based test environment and it's not required anymore. The UI and API tests have been separated for better organization and manageability. A new Cucumber scenario for end-to-end tests was added in EndtoEnd.feature. This will help to perform a complete test of the system, ensuring that all the components are interacting as expected. The User Logout scenario was removed from CrudUI.feature, as it had been moved to a more appropriate feature file regarding the user authentication. Finally, the Api.feature file was created to separate the API test scenarios from the UI, making the tests more readable and easy to maintain.