"Refactor test structure and implement new GraphQL approach
Deleted JobAPI.java and replaced it with a new API.java file that uses GraphQL for better and more efficient manipulation of data, replacing old RESTful approach. Updated tags in CukesRunner to reflect these changes and focus on the new GraphQL tests. This aims to improve test speed and reliability by reducing over-fetching and under-fetching of data."