"Refactor authentication code for API tests
Moved the user authentication code from `setUpApi()` method in the Hooks.java to `i_am_an_authenticated_user()` method in API.java. Now, the authentication token is generated in the Given step of each API test instead of in the setup step. This change makes it easier to isolate and debug authentication issues for individual tests, as each test is now responsible for its own setup."